+++ src/Bio/Graph/Constructions.hs
@@ -22,7 +22,6 @@ module Bio.Graph.Constructions
   , GraphState
   , PhylogeneticDAG(..)
   , PhylogeneticDAG2(..)
-  , PhylogeneticDAGish(..)
   , PostOrderDecorationDAG
   , SearchState
   , TopologicalResult
@@ -38,7 +37,6 @@ module Bio.Graph.Constructions
   , UnReifiedCharacterDAG
   ) where
 
-
 import Bio.Character
 import Bio.Character.Decoration.Additive
 import Bio.Character.Decoration.Continuous
@@ -46,13 +44,12 @@ import Bio.Character.Decoration.Discrete
 import Bio.Character.Decoration.Dynamic
 import Bio.Character.Decoration.Fitch
 import Bio.Character.Decoration.Metric
-import Bio.Graph.PhylogeneticDAG.Class
 import Bio.Graph.PhylogeneticDAG.Internal
 import Bio.Graph.ReferenceDAG.Internal
 import Bio.Graph.Solution
 import Bio.Sequence
-import Bio.Sequence.Metadata
 import Control.Evaluation
+import Data.Compact
 import Data.EdgeLength
 import Data.List.NonEmpty
 import Data.NodeLabel
@@ -87,7 +84,7 @@ type SearchState = EvaluationT IO GraphState
 -- |
 -- The state of the graph that partitions the evaluation model on one of two
 -- paths depending on the presence or absence of character states in the search.
-type GraphState = Either TopologicalResult DecoratedCharacterResult
+type GraphState = Compact (Either TopologicalResult DecoratedCharacterResult)
 
 
 -- |
@@ -139,6 +136,8 @@ type  UnifiedBlock =
     )
 
 
+-- |
+-- A character block resulting fro the READ command.
 type  UnifiedCharacterBlock
     = CharacterBlock
         UnifiedContinuousCharacter
@@ -149,9 +148,9 @@ type  UnifiedCharacterBlock
         UnifiedDynamicCharacter
 
 
-type  UnifiedMetadataBlock
-    = MetadataBlock
-        ()
+-- |
+-- A metadata block resulting fro the READ command.
+type  UnifiedMetadataBlock = MetadataBlock ()
 
 
 -- |
@@ -168,11 +167,8 @@ type  UnifiedSequences =
     )
 
 
-type  UnifiedMetadataSequence
-    = MetadataSequence
-        ()
-
-
+-- |
+-- A character sequence resulting fro the READ command.
 type  UnifiedCharacterSequence
     = CharacterSequence
         UnifiedContinuousCharacter
@@ -181,6 +177,11 @@ type  UnifiedCharacterSequence
         UnifiedDiscreteCharacter
         UnifiedDiscreteCharacter
         UnifiedDynamicCharacter
+
+
+-- |
+-- A metadata sequence resulting fro the READ command.
+type  UnifiedMetadataSequence = MetadataSequence ()
 
 
 -- |