+++ src/Bio/Graph/Node/Internal.hs
@@ -17,6 +17,7 @@
 {-# LANGUAGE FlexibleInstances      #-}
 {-# LANGUAGE FunctionalDependencies #-}
 {-# LANGUAGE MultiParamTypeClasses  #-}
+{-# LANGUAGE DerivingStrategies         #-}
 
 module Bio.Graph.Node.Internal
   ( EdgeSet
@@ -50,6 +51,7 @@ import Data.BitVector.LittleEndian
 import Data.EdgeSet
 import Data.Foldable
 import Data.List.NonEmpty          (NonEmpty (..))
+import Data.Text.Lazy              (Text, pack)
 import Data.TopologyRepresentation
 import Data.UnionSet
 import GHC.Generics
@@ -125,8 +127,11 @@ type ResolutionCache s = NonEmpty (ResolutionInformation s)
 -- |
 -- A newick representation of a subtree. 'Semigroup' instance used for subtree
 -- joining.
-newtype NewickSerialization = NS String
-  deriving (Eq, Generic, Ord)
+newtype NewickSerialization = NS Text
+  deriving newtype Eq
+  deriving         Generic
+  deriving newtype Ord
+  deriving newtype Show
 
 -- |
 -- A 'Lens' for the 'totalSubtreeCost' field in 'ResolutionInformation'
@@ -264,11 +269,6 @@ instance Semigroup NewickSerialization where
     (NS lhs) <> (NS rhs) = NS $ "(" <> lhs <> "," <> rhs <> ")"
 
 
-instance Show NewickSerialization where
-
-    show (NS s) = s
-
-
 instance (Show n, Show s) => Show (PhylogeneticNode2 s n) where
 
     show node = unlines
@@ -351,7 +351,7 @@ pNode2 = flip PNode2
 -- Construct a singleton newick string with a unique identifier that can be
 -- rendered to a string through its 'Show' instance.
 singletonNewickSerialization :: Show i => i -> NewickSerialization
-singletonNewickSerialization i = NS $ show i
+singletonNewickSerialization = NS . pack . show
 
 
 -- |