+++ src/Bio/Graph/PhylogeneticDAG/Internal.hs
@@ -38,6 +38,7 @@ module Bio.Graph.PhylogeneticDAG.Internal
   , HasMinimalNetworkContext(..)
   , setDefaultMetadata
   ) where
+
 import           Bio.Character.Decoration.Shared
 import           Bio.Character.Encodable
 import           Bio.Graph.LeafSet
@@ -48,11 +49,6 @@ import           Bio.Metadata.Discrete
 import           Bio.Metadata.DiscreteWithTCM
 import           Bio.Metadata.Dynamic
 import           Bio.Sequence
-import           Bio.Sequence.Block.Character    (CharacterBlock (..))
-import           Bio.Sequence.Block.Internal
-import           Bio.Sequence.Block.Metadata     (MetadataBlock (..))
-import           Bio.Sequence.Metadata           (MetadataSequence, getBlockMetadata)
-import qualified Bio.Sequence.Metadata           as M
 import           Control.Applicative             (liftA2)
 import           Control.Arrow                   ((***))
 import           Control.DeepSeq
@@ -60,7 +56,7 @@ import           Control.Lens                    as Lens
 import           Data.Bits
 import           Data.Foldable
 import           Data.Foldable.Custom            (sum')
-import           Data.GraphViz.Printing          hiding ((<>))
+import           Data.GraphViz.Printing
 import           Data.GraphViz.Types
 import           Data.HashMap.Lazy               (HashMap)
 import qualified Data.IntMap                     as IM
@@ -80,6 +76,7 @@ import           GHC.Generics
 import           Prelude                         hiding (zip)
 import           Text.Newick.Class
 import           Text.XML
+import           Type.Reflection                 (Typeable)
 
 
 -- |
@@ -143,7 +140,7 @@ data  PhylogeneticDAG2 m e n u v w x y z
                               e
                               (PhylogeneticNode2 (CharacterSequence u v w x y z) n)
     , columnMetadata     :: MetadataSequence m
-    } deriving (Generic)
+    } deriving (Generic, Typeable)
 
 
 -- |
@@ -455,7 +452,7 @@ renderSummary pdag@(PDAG2 dag _) = unlines
 
 
 renderMetadata :: Show m => MetadataSequence m -> String
-renderMetadata = unlines . fmap (show . getBlockMetadata) . toList . M.toBlocks
+renderMetadata = unlines . fmap (show . (^. blockMetadata)) . toList . (^. blockSequence)
 
 
 -- |
@@ -511,16 +508,15 @@ renderBlockSummary
   -> String
 renderBlockSummary (PDAG2 dag meta) key (costOfRooting, costOfNetworking, displayMay, block) = mconcat . (renderedPrefix:) .
     (renderBlockMeta pair :) $
-    [ unlines . fmap renderStaticCharacterSummary  . toList . uncurry zip . ( continuousBins ***  continuousBins)
-    , unlines . fmap renderStaticCharacterSummary  . toList . uncurry zip . (nonAdditiveBins *** nonAdditiveBins)
-    , unlines . fmap renderStaticCharacterSummary  . toList . uncurry zip . (   additiveBins ***    additiveBins)
-    , unlines . fmap renderStaticCharacterSummary  . toList . uncurry zip . (     metricBins ***      metricBins)
-    , unlines . fmap renderStaticCharacterSummary  . toList . uncurry zip . (  nonMetricBins ***   nonMetricBins)
-    , unlines . fmap renderDynamicCharacterSummary . toList . uncurry zip . (    dynamicBins ***     dynamicBins)
-    ] <*> [(mBlock, cBlock)]
+    [ unlines . fmap renderStaticCharacterSummary              . toList . uncurry zip . ((^.  continuousBin) *** (^.  continuousBin))
+    , unlines . fmap renderStaticCharacterWithAlphabetSummary  . toList . uncurry zip . ((^. nonAdditiveBin) *** (^. nonAdditiveBin))
+    , unlines . fmap renderStaticCharacterWithAlphabetSummary  . toList . uncurry zip . ((^.    additiveBin) *** (^.    additiveBin))
+    , unlines . fmap renderStaticCharacterWithAlphabetSummary  . toList . uncurry zip . ((^.      metricBin) *** (^.      metricBin))
+    , unlines . fmap renderStaticCharacterWithAlphabetSummary  . toList . uncurry zip . ((^.   nonMetricBin) *** (^.   nonMetricBin))
+    , unlines . fmap renderDynamicCharacterSummary             . toList . uncurry zip . ((^.     dynamicBin) *** (^.     dynamicBin))
+    ] <*> [(mBlock, block)]
   where
-    pair = (M.toBlocks meta ! key, block)
-    (MB mBlock, CB cBlock) = pair
+    pair@(mBlock, _) = ((meta ^. blockSequence) ! key, block)
 
     renderedPrefix = "Block " <> show key <> "\n\n"
 
@@ -541,9 +537,16 @@ renderBlockSummary (PDAG2 dag meta) key (costOfRooting, costOfNetworking, displa
             ]
 
     renderStaticCharacterSummary (m, c) = unlines
-        [ "    Name:   " <> show (m ^. characterName)
-        , "    Weight: " <> show (m ^. characterWeight)
-        , "    Cost:   " <> show (c ^. characterCost)
+        [ "    Name:     " <> show (m ^. characterName)
+        , "    Weight:   " <> show (m ^. characterWeight)
+        , "    Cost:     " <> show (c ^. characterCost)
+        ]
+
+    renderStaticCharacterWithAlphabetSummary (m, c) = unlines
+        [ "    Name:     " <> show (m ^. characterName)
+        , "    Weight:   " <> show (m ^. characterWeight)
+        , "    Cost:     " <> show (c ^. characterCost)
+        , "    "           <> show (m ^. characterAlphabet)
         ]
 
     renderDynamicCharacterSummary (m, c) = unlines