+++ src/Control/Parallel/Custom.hs
@@ -30,16 +30,16 @@ import Prelude                     hiding (zipWith)
 -- Map a function over a traversable structure in parrallel.
 -- Should be prefered over 'parMap' which is fixed to lists.
 parmap :: Traversable t => Strategy b -> (a -> b) -> t a -> t b
-parmap strat f = withStrategy (parTraversable strat) . fmap f
+parmap _strat f = withStrategy rpar . fmap f
 
 
 -- |
 -- Zip two traversable, zippable structures in parrallel with a function.
-parZipWith :: (Traversable t, Zip t) => Strategy c -> (a -> b -> c) -> t a -> t b -> t c
-parZipWith strat f lhs rhs = withStrategy (parTraversable strat) $ zipWith f lhs rhs
+parZipWith :: Zip t => Strategy c -> (a -> b -> c) -> t a -> t b -> t c
+parZipWith _strat f lhs rhs = withStrategy rpar $ zipWith f lhs rhs
 
 
 -- |
 -- Zip three traversable, zippable structures in parrallel with a function.
-parZipWith3 :: (Traversable t, Zip t) => Strategy d -> (a -> b -> c -> d) -> t a -> t b -> t c -> t d
-parZipWith3 strat f x y z = withStrategy (parTraversable strat) $ zap (zipWith f x y) z
+parZipWith3 :: Zip t => Strategy d -> (a -> b -> c -> d) -> t a -> t b -> t c -> t d
+parZipWith3 _strat f x y z = withStrategy rpar $ zap (zipWith f x y) z